powerci: add a script to get all the bundles from LAVA
this will allow to post stuff even if the runner broke,
or post stuff without polling involved.
or rebuilt the complete POWERCI db, based on the LAVA db.

export all results matching a string in the content name

prepare to using both /test amd /boot APIs to generate
the matching json to post to KcI or PcI

Signed-off-by: Marc Titinger <mtitinger@baylibre.com>